+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Bastardidentro.it channel
 
 ## [2.0.6] - 2020-02-21
 ### Added